crud-app/requirements.txt
flask
pytest
prometheus-flask-exporter
python-logging-loki
# opentelemetry-api
# opentelemetry-sdk
# opentelemetry-instrumentation-flask
# opentelemetry-exporter-jaeger

# python-dotenv
# opentelemetry-distro
# opentelemetry-exporter-otlp
# opentelemetry-bootstrap
